blogの位置をhomeに変更
