+++ blog/first/index.mdx
@@ -1,7 +1,7 @@
 ---
 title: "初投稿です"
 date: "2021-08-12"
-hero_image: "./cat.jpg"
+hero_image: "./cat_640px.jpg"
 ---
 
 Gatsby のチュートリアルで作成したものをそのまま使用しています
