+++ src/components/layout.js
@@ -37,11 +37,6 @@ const Layout = ({ pageTitle, children }) => {
               About
             </Link>
           </li>
-          <li className={navLinkItem}>
-            <Link to="/blog" className={navLinkText}>
-              Blog
-            </Link>
-          </li>
         </ul>
       </nav>
       <h1 className={heading}>{pageTitle}</h1>
