+++ src/pages/index.js
@@ -1,17 +1,35 @@
-import * as React from 'react'
-import Layout from '../components/layout'
-import { StaticImage } from 'gatsby-plugin-image'
+import * as React from "react";
+import { Link, graphql } from "gatsby";
+import Layout from "../components/layout";
 
-const IndexPage = () => {
+const BlogPage = ({ data }) => {
   return (
-    <Layout pageTitle="Home Page">
-      <p>I'm making this by following the Gatsby Tutorial.</p>
-      <StaticImage
-        alt="the working cat"
-        src="../images/shigotoneko.jpeg"
-      />
+    <Layout pageTitle="kizi ichiran">
+      {data.allMdx.nodes.map((node) => (
+        <article key={node.id}>
+          <h2>
+            <Link to={`/${node.slug}`}>{node.frontmatter.title}</Link>
+          </h2>
+          <p>Posted: {node.frontmatter.date}</p>
+        </article>
+      ))}
     </Layout>
-  )
-}
+  );
+};
 
-export default IndexPage
+export const query = graphql`
+  query {
+    allMdx(sort: { fields: frontmatter___date, order: DESC }) {
+      nodes {
+        frontmatter {
+          date(formatString: "MMMM D, YYYY")
+          title
+        }
+        id
+        slug
+      }
+    }
+  }
+`;
+
+export default BlogPage;
